Added PopulateMember(DataRow) function, has not been tested yet.
Made a function that placed the data from a DataRow object to the Member
object. Basically just pass it a DataRow from the DataTable that has a
member's information and the function returns a member object. I haven't
tested it yet so its just me typing out the syntax so I can save the
whole thing in a commit.