Created default Member constructor, DisplayMember() and PopulateMember()
So there is now a default constructor for member. If you create a new
Member without passing anything it'll be this default. Assigning it data
from SQL will obviously overwrite it but it can be useful when you don't
want to screw with SQL.

DisplayMember() is called just like testMember.DisplayMember() and it
will pop up a window with all their information, just use it when you
easily want to check what's in a member object. I'll make some of these
for the other classes as well.

Populate member is called the same way but you pass it a certain row
from a DataTable and it will populate the member with that row's data. I
put it in the LoginButton Click and it works so you can see what I'm
talking about there.

I put a lot of comments in my code, please read them when you see them
and let me know what you think. Happy Thanksgiving yall