created aggregate user account
moved all the relevant data for a member into a single class.
Account/Member/Login information is linked together and is now in this
one class. This class hold the methods to create a new user info with a
linked account/ member/login. Holds the method to create the active user
dataset with all relevant information from these three tables mentioned
above